update emulator pass
