+++ src/braket/emulation/passes/__init__.py
@@ -11,5 +11,7 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-from braket.emulation.passes.transformation_pass import TransformationPass  # noqa: F401
-from braket.emulation.passes.validation_pass import ValidationPass  # noqa: F401
+from braket.emulation.passes.passes import (  # noqa: F401
+    TransformationPass,
+    ValidationPass,
+)
