+++ src/braket/emulation/passes/passes.py
@@ -18,7 +18,60 @@ from abc import ABC, abstractmethod
 from braket.tasks.quantum_task import TaskSpecification
 
 
-class ValidationPass(ABC):
+class _EmulatorPass(ABC):
+    @abstractmethod
+    def run(self, task_specification: TaskSpecification) -> TaskSpecification:
+        raise NotImplementedError
+
+
+class TransformationPass(_EmulatorPass):
+    _supported_specifications: TaskSpecification
+
+    @abstractmethod
+    def transform(self, task_specification: TaskSpecification) -> TaskSpecification:
+        """
+        An emulator modifier is used to perform some potentially modifying validation
+        pass on an input program. Implementations of modify should return the same
+        specification if the input program passes validation and raise an error otherwise.
+
+        Args:
+            task_specification (TaskSpecification): The program to be evaluated against this
+                criteria.
+
+        Returns:
+            task_specificaiton (TaskSpecification): The (potentially) modified program
+        """
+        raise NotImplementedError
+
+    def run(self, task_specification: TaskSpecification) -> TaskSpecification:
+        """
+        Validate the input program and return the program, unmodified.
+
+        Args:
+            task_specification (TaskSpecification): The program to validate.
+
+        Returns:
+            TaskSpecification: The unmodified program passed in as input.
+        """
+        if isinstance(task_specification, self.supported_specifications):
+            task_specification = self.transform(task_specification)
+        return task_specification
+
+    def __call__(self, task_specification: TaskSpecification) -> TaskSpecification:
+        return self.run(task_specification)
+
+    @property
+    def supported_specifications(self) -> TaskSpecification:
+        """List of supported specifications for a ValidationPass
+
+        Returns:
+            TaskSpecification:
+                tuple, single, or union of supported specifications
+        """
+        return self._supported_specifications
+
+
+class ValidationPass(_EmulatorPass):
     _supported_specifications: TaskSpecification
 
     @abstractmethod
